Moved tests to the correct package
