+++ tygron-environment/src/test/java/nl/tudelft/contextproject/util/PolygonUtilTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.tygron;
+package nl.tudelft.contextproject.util;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
